Code fixes (#200)
* Code improvements

code improvements

* Code fixes

Fixing incorrect string comparisons

* Code review comment

Code review comment, fixing a failure test case.

* Fixing test failures

* Code review comment

Null check for record.preference()

+++ azure-mgmt-servicebus/src/test/java/com/microsoft/azure/management/servicebus/ServiceBusOperationsTests.java
@@ -17,6 +17,7 @@ import com.microsoft.rest.RestClient;
 import org.joda.time.Period;
 import org.junit.Assert;
 import org.junit.Test;
+
 import java.util.concurrent.TimeUnit;
 
 public class ServiceBusOperationsTests extends TestBase {
@@ -46,7 +47,7 @@ public class ServiceBusOperationsTests extends TestBase {
 
     @Override
     protected void cleanUpResources() {
-        if (RG_NAME != null) {
+        if (RG_NAME != null && resourceManager != null) {
             resourceManager.resourceGroups().deleteByName(RG_NAME);
         }
     }
